requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5        # via requests
chardet==3.0.4            # via readability-lxml, requests
click==7.1.2              # via -r requirements.in
cssselect==1.1.0          # via readability-lxml
gensim==3.8.3             # via -r requirements.in
html2text==2020.1.16      # via -r requirements.in
idna==2.10                # via requests
joblib==1.0.0             # via scikit-learn
lxml==4.6.2               # via readability-lxml
mecab-python3==1.0.3      # via -r requirements.in
numpy==1.19.4             # via gensim, scikit-learn, scipy
readability-lxml==0.8.1   # via -r requirements.in
requests==2.25.0          # via -r requirements.in
scikit-learn==0.23.2      # via -r requirements.in
scipy==1.5.4              # via gensim, scikit-learn
six==1.15.0               # via gensim
smart-open==4.0.1         # via gensim
threadpoolctl==2.1.0      # via scikit-learn
urllib3==1.26.2           # via requests
